@@ -13,48 +13,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/forms": "^8.0.0",
+    "@angular/common": "^11.2.11",
+    "@angular/core": "^11.2.11",
+    "@angular/forms": "^11.2.11",
     "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@angular/router": "^8.0.0",
-    "@ionic-native/core": "^5.7.0",
-    "@ionic-native/splash-screen": "^5.7.0",
-    "@ionic-native/status-bar": "^5.7.0",
-    "@ionic/angular": "^4.4.2",
-    "@ionic/lab": "^2.0.0",
+    "@angular/platform-browser": "^11.2.11",
+    "@angular/platform-browser-dynamic": "^11.2.11",
+    "@angular/router": "^11.2.11",
+    "@ionic-native/core": "^5.32.1",
+    "@ionic-native/splash-screen": "^5.32.1",
+    "@ionic-native/status-bar": "^5.32.1",
+    "@ionic/angular": "^5.6.4",
+    "@ionic/lab": "^3.2.10",
     "@ionic/pro": "2.0.4",
-    "core-js": "^3.1.3",
-    "rxjs": "~6.5.2",
-    "zone.js": "~0.9.1"
+    "core-js": "^3.10.2",
+    "rxjs": "~6.6.7",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.800.1",
-    "@angular-devkit/build-angular": "~0.800.1",
-    "@angular-devkit/core": "~8.0.1",
-    "@angular-devkit/schematics": "~8.0.1",
-    "@angular/cli": "~8.0.1",
-    "@angular/compiler": "~8.0.0",
-    "@angular/compiler-cli": "~8.0.0",
-    "@angular/language-service": "~8.0.0",
-    "@ionic/angular-toolkit": "~1.5.1",
-    "@types/jasmine": "~3.3.13",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~12.0.4",
-    "codelyzer": "~5.1.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.2.0",
-    "tslint": "~5.17.0",
-    "typescript": "3.4.1"
+    "@angular-devkit/architect": "~0.1102.9",
+    "@angular-devkit/build-angular": "~0.1102.9",
+    "@angular-devkit/core": "~11.2.9",
+    "@angular-devkit/schematics": "~11.2.9",
+    "@angular/cli": "~11.2.9",
+    "@angular/compiler": "~11.2.11",
+    "@angular/compiler-cli": "~11.2.11",
+    "@angular/language-service": "~11.2.11",
+    "@ionic/angular-toolkit": "~3.1.1",
+    "@types/jasmine": "~3.6.9",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~14.14.41",
+    "codelyzer": "~6.0.1",
+    "jasmine-core": "~3.7.1",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.3.2",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "~7.0.0",
+    "ts-node": "~9.1.1",
+    "tslint": "~5.20.1",
+    "typescript": "4.0.5"
   },
   "description": "An Ionic project"
 }
