@@ -1,26 +1,23 @@
-import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Component, OnInit } from "@angular/core";
+import { Observable } from "rxjs";
 
-import { InfoService, SearchType } from './../../services/info.service';
+import { InfoService, SearchType } from "./../../services/info.service";
 
 @Component({
-  selector: 'app-info',
-  templateUrl: './info.page.html',
-  styleUrls: ['./info.page.scss'],
+  selector: "app-info",
+  templateUrl: "./info.page.html",
+  styleUrls: ["./info.page.scss"],
 })
 export class InfoPage implements OnInit {
-
   results: Observable<any>;
-  searchTerm: '';
+  searchTerm: "";
   type: SearchType = SearchType.all;
 
-  constructor(private infoService: InfoService) { }
+  constructor(private infoService: InfoService) {}
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
-  searchChanged() {
-    // Call our service function which returns an Observable
+  searchChanged(): void {
     this.results = this.infoService.searchData(this.searchTerm, this.type);
   }
 }
